// packages/cli/src/components/command-menu/types.ts
import type { DialogContextValue } from "../../providers/dialog";
import type { DialogConfig } from "../../providers/dialog/types";
import type { ToastContextValue } from "../../providers/toast";

export type CommandContext = {
    exit: () => void;
    toast: ToastContextValue;
    dialog: DialogContextValue
};

export type Command = {
    name: string;
    description: string;
    value: string;
    action?: (ctx: CommandContext) => void;
};